Add proper file exist check

+++ graphion/api/filter/distribution.py
@@ -1,5 +1,5 @@
 """
-Author(s): Steven van den Broek
+Author(s): Steven van den Broek, Tom Udding
 Created: 2019-05-18
 Edited: 2019-05-22
 """
@@ -9,7 +9,7 @@ from json import dump, dumps
 from graphion.filtering.degree_selection import generate_selection
 from graphion.graphing.generator import getFilePath
 import time
-import os
+from os.path import exists
 
 apiDegreeBlueprint = Blueprint('apiMatrixBlueprint', __name__, template_folder='templates')
 
@@ -19,11 +19,10 @@ apiDegreeBlueprint = Blueprint('apiMatrixBlueprint', __name__, template_folder='
 @apiDegreeBlueprint.route('/api/filter/distribution/<type>/', methods=['GET'], strict_slashes=False)
 def degreeAPI(file=None, type=None, dir=None):
     filePath = 'api/filter/cached_plots/{}-{}-{}.json'.format(file, type, dir)
-    try:
+    if (exists(filePath)):
         with open(filePath, 'r') as json_file:
             return json_file.read()
-    except FileNotFoundError:
-
+    else:
         with open(filePath, 'w+') as json_file:
             plot = generate_selection(getFilePath(file), kind=type, dir=dir)
             start = time.time()
